requirements.txt
datasets==3.6.0
peft==0.17.0
python-dotenv==1.1.1
torch==2.7.1
transformers==4.52.4
